pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gambiarra-client"
version = "1.0.0"
description = "Cliente Python para Gambiarra LLM Club Arena"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    { name = "Gambiarra Club", email = "contact@gambiarra.club" }
]
dependencies = [
    "websockets>=12.0",
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
gambiarra-client = "gambiarra_client.cli:run"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["gambiarra_client*"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.ruff]
line-length = 100
target-version = "py38"
